+++ backend/tools/Faktur.Relocation.Worker/Program.cs
@@ -1,0 +1,15 @@
+﻿namespace Faktur.Relocation.Worker;
+
+internal class Program
+{
+  private static void Main(string[] args)
+  {
+    HostApplicationBuilder builder = Host.CreateApplicationBuilder(args);
+
+    Startup startup = new(builder.Configuration);
+    startup.ConfigureServices(builder.Services);
+
+    IHost host = builder.Build();
+    host.Run();
+  }
+}